django_test_oauth/requirements.txt
django==4.2.7
django-allauth==0.58.2